pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "threat-log-analyzer"
version = "0.2.0"
description = "Analyze SSH/auth logs for suspicious activity (educational & defensive)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["PyYAML>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
tla = "threat_log_analyzer.cli:main"

[tool.setuptools.package-data]
threat_log_analyzer = ["rules/*.yaml"]